tests/requirements.txt
pytest==7.2.0
pytest-cov==4.0.0
black==22.10.0
python-dotenv==0.21.0
beautifulsoup4
blinker>=1.6